+++ python/llm/dev/test/lint-python
@@ -21,7 +21,9 @@ SCRIPT_DIR="$( cd "$( dirname "$0" )" && pwd )"
 PYTHON_ROOT_DIR="$SCRIPT_DIR/.."
 echo $PYTHON_ROOT_DIR
 PATHS_TO_CHECK="$SCRIPT_DIR/../../src"
-PATTERNS_TO_EXCLUDE="__init__.py,log4Error.py,$SCRIPT_DIR/../../src/bigdl/llm/langchain/*"
+PATTERNS_TO_EXCLUDE=(
+    "__init__.py,log4Error.py,$SCRIPT_DIR/../../src/bigdl/llm/langchain/*,"
+    "$SCRIPT_DIR/../../src/bigdl/llm/transformers/training_args.py")
 PEP8_REPORT_PATH="$PYTHON_ROOT_DIR/test/pep8-report.txt"
 PYLINT_REPORT_PATH="$PYTHON_ROOT_DIR/test/pylint-report.txt"
 PYLINT_INSTALL_INFO="$PYTHON_ROOT_DIR/test/pylint-info.txt"
